Add logic to Map button to pass location to external program Add button to show calendar Add logic to Calendar button to open external calendar
