+++ app/src/main/java/com/lambda_labs/community_calendar/view/HomeFragment.kt
@@ -2,10 +2,7 @@ package com.lambda_labs.community_calendar.view
 
 import EventsQuery
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
-import android.provider.CalendarContract
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -152,69 +149,14 @@ class HomeFragment : Fragment() {
 
         // Tomorrow tab filters by tomorrows events
         txt_events_tomorrow.setOnClickListener {
-
-            // Pass event location to external maps app
-            // Get the values from the selected event details
-            val eventID=4
-            val encodedLatitude= Uri.encode(events[eventID].locations()?.get(0)?.latitude().toString())
-            val encodedLongitude= Uri.encode(events[eventID].locations()?.get(0)?.longitude().toString())
-            val encodedLocationName=Uri.encode(events[eventID].locations()?.get(0)?.name())
-
-            // By Name
-            //val uriMaps: Uri = Uri.parse("geo:0,0?z=17&q=$encodedLocationName")
-
-            // By Coordinates
-            val uriMaps: Uri = Uri.parse("geo:0,0?q=$encodedLatitude,$encodedLongitude($encodedLocationName)&z=17")
-
-            // Create the Intent object to have the OS decide how to show the place on the Map
-            val intent:Intent=Intent(Intent.ACTION_VIEW,uriMaps)
-            startActivity(intent)
-
-
-            /*changeColor(it)
+            changeColor(it)
             changeDay(getTomorrow())
-            isEmpty("tomorrow")*/
+            isEmpty("tomorrow")
         }
 
         // Weekend tab filters by events this weekend
         txt_events_this_weekend.setOnClickListener {
-
-            // Pass the event date and time to external calendar app
-            // Convert the times from a String to a Date to a Calendar and into epoch milliseconds
-            val eventID=4
-            // Start time
-            var timeToString: String =events[eventID].start().toString()
-            var stringToDate: Date = stringToDate(timeToString)
-            var dateToCalendar: Calendar = toCalendar(stringToDate)
-            val startTimeMilliseconds: Long =dateToCalendar.timeInMillis
-
-            // End time
-            timeToString =events[eventID].end().toString()
-            stringToDate = stringToDate(timeToString)
-            dateToCalendar = toCalendar(stringToDate)
-            val endTimeMilliseconds: Long =dateToCalendar.timeInMillis
-
-            // Extract the event information for the calendar entry
-            val eventName=events[eventID].title()
-            val eventDescription=events[eventID].description()
-            val eventLatitude=events[eventID].locations()?.get(0)?.latitude()
-            val eventLongitude=events[eventID].locations()?.get(0)?.longitude()
-            val eventCoordinates = "$eventLatitude,$eventLongitude"
-
-            // Create the Intent object with extras of event info to launch externally
-            val intent:Intent = Intent(Intent.ACTION_INSERT)
-                .setData(CalendarContract.Events.CONTENT_URI)
-                .putExtra(CalendarContract.EXTRA_EVENT_BEGIN_TIME, startTimeMilliseconds)
-                .putExtra(CalendarContract.EXTRA_EVENT_END_TIME, endTimeMilliseconds)
-                .putExtra(CalendarContract.Events.TITLE, eventName)
-                .putExtra(CalendarContract.Events.DESCRIPTION, eventDescription)
-                .putExtra(CalendarContract.Events.EVENT_LOCATION, eventCoordinates)
-                .putExtra(CalendarContract.Events.AVAILABILITY, CalendarContract.Events.AVAILABILITY_BUSY)
-            startActivity(intent)
-
-
-
-            /*changeColor(it)
+            changeColor(it)
             filterList.clear()
             getWeekendDates().forEach { date ->
                 filterList.addAll(events.filter {
@@ -225,7 +167,7 @@ class HomeFragment : Fragment() {
             main_event_recycler.adapter?.notifyDataSetChanged()
             val displayWeekend =
                 "${getDisplayDay(getWeekendDates()[0])} - ${getDisplayDay(getWeekendDates()[2])}"
-            txt_event_date.text = displayWeekend*/
+            txt_event_date.text = displayWeekend
         }
 
         // All upcoming filters by dates after today
