+++ app/src/main/res/layout/fragment_event_details.xml
@@ -333,6 +333,21 @@
                 android:textAllCaps="false"
                 app:layout_constraintTop_toBottomOf="@+id/btn_attend"/>
 
+            <com.google.android.material.button.MaterialButton
+                android:id="@+id/btn_calendar"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="16dp"
+                android:layout_marginEnd="16dp"
+                android:visibility="gone"
+                style="@style/Widget.MaterialComponents.Button.OutlinedButton"
+                android:background="@drawable/attend_button_outline"
+                android:text="Save in Calendar"
+                android:textColor="#FFFFFF"
+                android:backgroundTint="#21242C"
+                android:textAllCaps="false"
+                app:layout_constraintTop_toBottomOf="@+id/btn_map"/>
+
         </androidx.constraintlayout.widget.ConstraintLayout>
     </ScrollView>
